MPS-22195 NPE on deleting devkit
DeleteModuleHelper: fix bug itself

Rewrite NewDevKitDialog:
 * UI layout, labels and focus
 * change path on name changing (until path set manually)
 * checking name and location on the fly

 DeleteModules_Action: html markup for label + mnemonic for checkbox

 ProjectPropertiesComponent: inline ModulePathChooser and set in project folder as default for 'add module' chooser

+++ workbench/mps-workbench/source/jetbrains/mps/workbench/actions/module/DeleteModuleHelper.java
@@ -85,9 +85,12 @@ public class DeleteModuleHelper {
 
       if (module instanceof AbstractModule) {
         AbstractModule curModule = (AbstractModule) module;
-        String outputPath = curModule.getOutputPath().toPath().toString();
-        deleteFile(outputPath);
-        deleteFile(FileGenerationUtil.getCachesPath(outputPath));
+        final IFile output = curModule.getOutputPath();
+        if (output != null) {
+          String outputPath = output.toPath().toString();
+          deleteFile(outputPath);
+          deleteFile(FileGenerationUtil.getCachesPath(outputPath));
+        }
 
         if (curModule.getDescriptorFile() != null) {
           curModule.getDescriptorFile().delete();
@@ -99,7 +102,7 @@ public class DeleteModuleHelper {
 
         if (curModule.getDescriptorFile() != null) {
           IFile moduleFolder = curModule.getDescriptorFile().getParent();
-          if (moduleFolder !=null && deleteDirIfEmpty(moduleFolder)) {
+          if (moduleFolder != null && deleteDirIfEmpty(moduleFolder)) {
             moduleFolder.delete();
           }
         }
